+++ src/components/shipment/ShipmentList.vue
@@ -26,32 +26,99 @@
       </router-link>
     </div>
 
-    <div v-if="filteredShipments.length === 0" class="text-center">
+    <div v-if="paginatedShipments.length === 0" class="text-center">
       <p>Aucune expédition trouvée.</p>
     </div>
 
     <table v-else class="table table-striped table-bordered">
       <thead>
         <tr>
-          <th scope="col">ID Expédition</th>
-          <th scope="col">Nom Destinataire</th>
-          <th scope="col">Prénom Destinataire</th>
-          <th scope="col">Téléphone Destinataire</th>
-          <th scope="col">Destination</th>
-          <th scope="col">Date Expédition</th>
+          <th @click="toggleSortOrder('id')" style="cursor: pointer">
+            ID 
+            <span v-if="sortBy === 'id'">
+              <i
+                :class="
+                  sortOrder === 'asc' ? 'fas fa-arrow-up' : 'fas fa-arrow-down'
+                "
+              ></i>
+            </span>
+          </th>
+          <th
+            @click="toggleSortOrder('nom_destinataire')"
+            style="cursor: pointer"
+          >
+            Nom Destinataire
+            <span v-if="sortBy === 'nom_destinataire'">
+              <i
+                :class="
+                  sortOrder === 'asc' ? 'fas fa-arrow-up' : 'fas fa-arrow-down'
+                "
+              ></i>
+            </span>
+          </th>
+          <th
+            @click="toggleSortOrder('prenom_destinataire')"
+            style="cursor: pointer"
+          >
+            Prénom Destinataire
+            <span v-if="sortBy === 'prenom_destinataire'">
+              <i
+                :class="
+                  sortOrder === 'asc' ? 'fas fa-arrow-up' : 'fas fa-arrow-down'
+                "
+              ></i>
+            </span>
+          </th>
+          <th @click="toggleSortOrder('destination')" style="cursor: pointer">
+            Téléphone
+            <span v-if="sortBy === 'destination'">
+              <i
+                :class="
+                  sortOrder === 'asc' ? 'fas fa-arrow-up' : 'fas fa-arrow-down'
+                "
+              ></i>
+            </span>
+          </th>
+          <th
+            @click="toggleSortOrder('destination')"
+            style="cursor: pointer"
+          >
+            Destination
+            <span v-if="sortBy === 'destination'">
+              <i
+                :class="
+                  sortOrder === 'asc' ? 'fas fa-arrow-up' : 'fas fa-arrow-down'
+                "
+              ></i>
+            </span>
+          </th>
+          <th
+            @click="toggleSortOrder('date_expedition')"
+            style="cursor: pointer"
+          >
+            Date Expédition
+            <span v-if="sortBy === 'date_expedition'">
+              <i
+                :class="
+                  sortOrder === 'asc' ? 'fas fa-arrow-up' : 'fas fa-arrow-down'
+                "
+              ></i>
+            </span>
+          </th>
+
           <th scope="col">Utilisateur</th>
           <th scope="col">Colis</th>
           <th scope="col" class="text-center">Actions</th>
         </tr>
       </thead>
       <tbody>
-        <tr v-for="shipment in filteredShipments" :key="shipment.id">
+        <tr v-for="shipment in paginatedShipments" :key="shipment.id">
           <td>{{ shipment.id }}</td>
           <td>{{ shipment.nom_destinataire }}</td>
           <td>{{ shipment.prenom_destinataire }}</td>
           <td>{{ shipment.telephone_destinataire }}</td>
           <td>{{ shipment.destination }}</td>
-          <td>{{ shipment.date_expedition }}</td>
+          <td>{{ new Date(shipment.date_expedition).toLocaleDateString() }}</td>
           <td>
             {{
               shipment.utilisateur ? shipment.utilisateur.nom : 'Non attribué'
@@ -83,6 +150,25 @@
         </tr>
       </tbody>
     </table>
+
+
+    <div class="d-flex justify-content-center mt-4">
+      <button
+        class="btn btn-outline-secondary"
+        :disabled="currentPage === 1"
+        @click="prevPage"
+      >
+        <i class="fas fa-chevron-left"></i> Précédent
+      </button>
+      <span class="mx-3">Page {{ currentPage }} sur {{ totalPages }}</span>
+      <button
+        class="btn btn-outline-secondary"
+        :disabled="currentPage === totalPages"
+        @click="nextPage"
+      >
+        Suivant <i class="fas fa-chevron-right"></i>
+      </button>
+    </div>
   </div>
 </template>
 
@@ -99,9 +185,12 @@ const parcelStore = useParcelStore()
 const userStore = useUserStore()
 
 const shipments = ref([])
-const parcels = ref([])
-const users = ref([])
 const searchQuery = ref('')
+const sortBy = ref('id') 
+const sortOrder = ref('asc') 
+
+const currentPage = ref(1)
+const itemsPerPage = 4
 
 const fetchAndMapShipments = async () => {
   try {
@@ -116,9 +205,6 @@ const fetchAndMapShipments = async () => {
       ),
       colis: parcelStore.parcels.find(parcel => parcel.id === shipment.colisId),
     }))
-
-    parcels.value = parcelStore.parcels
-    users.value = userStore.users
   } catch (error) {
     console.error('Erreur lors du chargement des données :', error)
     toast.error('Erreur lors du chargement des données.')
@@ -127,12 +213,17 @@ const fetchAndMapShipments = async () => {
 
 onMounted(fetchAndMapShipments)
 
-const selectedDestination = ref('')
-const filteredShipments = computed(() =>
-  shipments.value.filter(shipment => {
-    const matchesDestination = selectedDestination.value
-      ? shipment.destination === selectedDestination.value
-      : true
+const toggleSortOrder = field => {
+  if (sortBy.value === field) {
+    sortOrder.value = sortOrder.value === 'asc' ? 'desc' : 'asc'
+  } else {
+    sortBy.value = field
+    sortOrder.value = 'asc'
+  }
+}
+
+const filteredShipments = computed(() => {
+  return shipments.value.filter(shipment => {
     const matchesSearch = searchQuery.value
       ? shipment.nom_destinataire
           .toLowerCase()
@@ -144,30 +235,55 @@ const filteredShipments = computed(() =>
           .toLowerCase()
           .includes(searchQuery.value.toLowerCase())
       : true
-    return matchesDestination && matchesSearch
+    return matchesSearch
   })
-)
+})
 
+const sortedShipments = computed(() => {
+  return [...filteredShipments.value].sort((a, b) => {
+    const fieldA = a[sortBy.value] || ''
+    const fieldB = b[sortBy.value] || ''
+    if (fieldA < fieldB) return sortOrder.value === 'asc' ? -1 : 1
+    if (fieldA > fieldB) return sortOrder.value === 'asc' ? 1 : -1
+    return 0
+  })
+})
+
+const paginatedShipments = computed(() => {
+  const startIndex = (currentPage.value - 1) * itemsPerPage
+  const endIndex = startIndex + itemsPerPage
+  return sortedShipments.value.slice(startIndex, endIndex)
+})
+
+const totalPages = computed(() => {
+  return Math.ceil(sortedShipments.value.length / itemsPerPage)
+})
+
+const nextPage = () => {
+  if (currentPage.value < totalPages.value) currentPage.value++
+}
+
+const prevPage = () => {
+  if (currentPage.value > 1) currentPage.value--
+}
 const deleteShipment = async id => {
-  const result = await shipmentStore.deleteShipment(id)
-  if (result.success) {
-    toast.success('Expédition supprimée avec succès.')
-    await fetchAndMapShipments()
+  const confirmation = confirm('Voulez-vous vraiment supprimer cette expédition ?')
+  if (confirmation) {
+    const result = await shipmentStore.deleteShipment(id)
+    if (result.success) {
+      toast.success('Expédition supprimée avec succès.')
+    } else {
+      toast.error('imposible de supprimer cette expedition est lié à colis')
+    }
   } else {
-    console.error(
-      "Erreur lors de la suppression de l'expédition :",
-      result.error
-    )
-    toast.error(
-      `Erreur lors de la suppression de l'expédition : ${result.error}`
-    )
+    toast.warning('Suppression annulée.')
   }
 }
 </script>
 
 <style scoped>
 .text-primary {
-  color: #4a4a4a;
+  color: #3fb59e;
 }
 
 .add-button:hover {
@@ -209,4 +325,3 @@ const deleteShipment = async id => {
   margin-block-start: 80px;
 }
 </style>
-
